requirements.txt
beautifulsoup4==4.12.0
certifi==2022.12.7
charset-normalizer==3.1.0
contourpy==1.0.7
cycler==0.11.0
distlib==0.3.6
filelock==3.10.7
fonttools==4.39.3
idna==3.4
imageio==2.27.0
imageio-ffmpeg==0.4.8
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.2
opencv-python==4.7.0.72
packaging==23.0
Pillow==9.4.0
platformdirs==3.2.0
psutil==5.9.4
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.28.2
six==1.16.0
soupsieve==2.4
urllib3==1.26.15
